Home page implementation
Nice linear gradient for ease of reading. Config partial has a function that returns text color based on a lightness of a given background color. The function is used now for setting a color of different elements on the page. Config also has a transition mixin which is used in a bunch of places of the page. Menu and Home partials show how to work with nesting.